requirements.txt
grpcio-tools
fastapi
uvicorn
pycryptodome
requests